+++ metals/src/main/scala/scala/meta/internal/metals/MetalsLanguageServer.scala
@@ -1607,7 +1607,9 @@ class MetalsLanguageServer(
   def rename(
       params: RenameParams
   ): CompletableFuture[WorkspaceEdit] =
-    CancelTokens.future { token => renameProvider.rename(params, token) }
+    CancelTokens.future { token =>
+      renameProvider.rename(params, compilers, token)
+    }
 
   @JsonRequest("textDocument/references")
   def references(
