+++ tests/unit/src/main/scala/tests/TestingServer.scala
@@ -1363,7 +1363,21 @@ final case class TestingServer(
         } else {
           val code = buffers.get(path).get
           if (renames.getDocumentChanges() == null) {
-            file -> code
+            if (renames.getChanges() == null)
+              file -> code
+            else {
+              val edited = for {
+                validLocations <-
+                  renames
+                    .getChanges()
+                    .asScala
+                    .get(path.toString())
+              } yield TextEdits.applyEdits(
+                code,
+                validLocations.asScala.toList,
+              )
+              file -> edited.getOrElse(code)
+            }
           } else {
             val renamed = renameFile(file, renames)
             renamed -> TestRanges
